+++ Assets/ConsentManagementProvider/Plugins/iOS/Source/Unity.mm
@@ -12,6 +12,25 @@ typedef void (*СallbackCharMessage) (const char*);
 
 extern "C"
 {
+    NSString* _getString(char * value)
+    {
+        return [NSString stringWithFormat:@"%s", value];
+    }
+    
+    char* _checkGetString(char * value)
+    {
+        NSLog(@"%s","Check string");
+        if (value == NULL)
+        {
+            NSLog(@"%s","Checked string is NULL");
+            return NULL;
+        }
+        char* res = (char*)malloc(strlen(value)+1);
+        strcpy (res, value);
+        NSLog(@"data = %s", value);
+        return res;
+    }
+
     void _setCallbackDefault (СallbackCharMessage callback){
         if (swiftBridge == nil)
             swiftBridge = [[SwiftBridge alloc] init];
@@ -68,7 +87,7 @@ extern "C"
 
     void _addTargetingParamForCampaignType(int campaignType, char* key, char* value)
     {
-        [swiftBridge addTargetingParamWithCampaignType:campaignType key:[NSString stringWithFormat:@"%s", key] value:[NSString stringWithFormat:@"%s", value]];
+        [swiftBridge addTargetingParamWithCampaignType:campaignType key:_getString(key) value:_getString(value)];
     }
 
     void _setTransitionCCPAAuth(bool value)
@@ -83,12 +102,17 @@ extern "C"
 
     void _configLib(int accountId, int propertyId, char* propertyName, bool gdpr, bool ccpa, bool usnat, SPMessageLanguage language, char* gdprPmId, char* ccpaPmId, char* usnatPmId)
     {
-        [swiftBridge configLibWithAccountId:accountId propertyId:propertyId propertyName:[NSString stringWithFormat:@"%s", propertyName] gdpr:gdpr ccpa:ccpa usnat:usnat language:language gdprPmId:[NSString stringWithFormat:@"%s", gdprPmId] ccpaPmId:[NSString stringWithFormat:@"%s", ccpaPmId] usnatPmId:[NSString stringWithFormat:@"%s", usnatPmId]];
+        [swiftBridge configLibWithAccountId:accountId propertyId:propertyId propertyName:_getString(propertyName) gdpr:gdpr ccpa:ccpa usnat:usnat language:language gdprPmId:_getString(gdprPmId) ccpaPmId:_getString(ccpaPmId) usnatPmId:_getString(usnatPmId)];
     }
 
-    void _loadMessage(char * authId)
+    void _loadMessage()
     {
-        [swiftBridge loadMessageWithAuthId:[NSString stringWithFormat:@"%s", authId]];
+        [swiftBridge loadMessageWithAuthId:nil];
+    }
+
+    void _loadMessageWithAuthId(char * authId)
+    {
+        [swiftBridge loadMessageWithAuthId:_getString(authId)];
     }
 
     void _loadGDPRPrivacyManager()
@@ -124,17 +148,17 @@ extern "C"
 
     void _addVendor(char* vendor)
     {
-        [swiftBridge addCustomVendorWithVendor:[NSString stringWithFormat:@"%s", vendor]];
+        [swiftBridge addCustomVendorWithVendor:_getString(vendor)];
     }
 
     void _addCategory(char* category)
     {
-        [swiftBridge addCustomCategoryWithCategory:[NSString stringWithFormat:@"%s", category]];
+        [swiftBridge addCustomCategoryWithCategory:_getString(category)];
     }
 
     void _addLegIntCategory(char* legIntCategory)
     {
-        [swiftBridge addCustomLegIntCategoryWithLegIntCategory:[NSString stringWithFormat:@"%s", legIntCategory]];
+        [swiftBridge addCustomLegIntCategoryWithLegIntCategory:_getString(legIntCategory)];
     }
 
     void _clearCustomArrays()