+++ Assets/ExampleApp/Scripts/PrivacySettings.cs
@@ -4,7 +4,7 @@ using UnityEngine.UI;
 using System.Collections.Generic;
 
 
-public class PrivacySettings : MonoBehaviour, IOnConsentReady
+public class PrivacySettings : MonoBehaviour, IOnConsentReady, IOnConsentSpFinished
 {
     public int accountId = 22;
     public int propertyId = 16893;    
@@ -34,6 +34,7 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
     }
 
     public Text consentValueText;
+    public Text authIdText;
     public Button loadMessageButton;
     public Button gdprPrivacySettingsButton;
     public Button ccpaPrivacySettingsButton;
@@ -41,12 +42,14 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
     public Button customConsentButton;
     public Button deleteCustomConsentButton;
     public Button clearDataButton;
+    public Text sdkStatus;
 
     private string storedConsentString = null;
 
     private void Awake()
     {
         ConsentMessenger.AddListener<IOnConsentReady>(gameObject);
+        ConsentMessenger.AddListener<IOnConsentSpFinished>(gameObject);
 
         List<SpCampaign> spCampaigns = new List<SpCampaign>();
         List<TargetingParam> gdprParams = new List<TargetingParam> { new TargetingParam("location", "EU") };
@@ -63,6 +66,8 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
         SpCampaign usnat = new SpCampaign(CAMPAIGN_TYPE.USNAT, usnatParams);
         spCampaigns.Add(usnat);
         campaignTypes.Add(CAMPAIGN_TYPE.USNAT);
+
+        UpdateSdkStatus("Running");
 
         CMP.Initialize(
             spCampaigns: spCampaigns,
@@ -88,6 +93,7 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
 
     public void OnLoadMessagePress()
     {
+        UpdateSdkStatus("Running");
         storedConsentString = null;
         updateUI();
         CMP.LoadMessage(authId: authId);
@@ -95,6 +101,7 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
 
     public void OnGDPRPrivacyManagerButtonClick()
     {
+        UpdateSdkStatus("Running");
         CMP.LoadPrivacyManager(
             campaignType: CAMPAIGN_TYPE.GDPR,
             pmId: gdprPmId,
@@ -104,6 +111,7 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
 
     public void OnCCPAPrivacyManagerButtonClick()
     {
+        UpdateSdkStatus("Running");
         CMP.LoadPrivacyManager(
             campaignType: CAMPAIGN_TYPE.CCPA,
             pmId: ccpaPmId,
@@ -113,6 +121,7 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
 
     public void OnUSNATPrivacyManagerButtonClick()
     {
+        UpdateSdkStatus("Running");
         CMP.LoadPrivacyManager(
             campaignType: CAMPAIGN_TYPE.USNAT,
             pmId: usnatPmId,
@@ -148,6 +157,7 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
 
     public void OnClearDataPress()
     {
+        UpdateSdkStatus("Not Started");
         CMP.ClearAllData();
         storedConsentString = null;
         updateUI();
@@ -163,6 +173,11 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
         updateUI();
     }
 
+    public void OnConsentSpFinished()
+    {
+        UpdateSdkStatus("Finished");
+    }
+
     private void updateUI()
     {
         if (storedConsentString != null)
@@ -175,6 +190,7 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
             deleteCustomConsentButton.interactable = true;
             clearDataButton.interactable = true;
             consentValueText.text = storedConsentString;
+            authIdText.text = CMP.GetBridgeString("AuthId:"+authId);
         }
         else
         {
@@ -186,12 +202,19 @@ public class PrivacySettings : MonoBehaviour, IOnConsentReady
             deleteCustomConsentButton.interactable = false;
             clearDataButton.interactable = false;
             consentValueText.text = "-";
+            authIdText.text = "-";
         }
+    }
+
+    void UpdateSdkStatus(string status)
+    {
+        sdkStatus.text = "SDK:"+status;
     }
 
     private void OnDestroy()
     {
         ConsentMessenger.RemoveListener<IOnConsentReady>(gameObject);
+        ConsentMessenger.RemoveListener<IOnConsentSpFinished>(gameObject);
         CMP.Dispose();
     }
 }