+++ UI-TESTS/UnitTest1.cs
@@ -322,7 +322,42 @@ namespace UnityAppiumTests
         	data = pages.nativeAppLayer.getConsentValueText();
 			Console.WriteLine($"ConsentValueText: {data}");
     		Assert.That(data!="-", Is.True);
+		}
 
+		[Test]
+		public void AuthIdTest()
+		{
+			Console.WriteLine(">>>AuthIdTest");
+			if (driver == null)
+			{
+				Assert.Fail("Driver has not been initialized.");
+			}
+
+			string firstLayerContext = pages.preFirstLayer.SelectFirstLayer();
+
+			pages.firstLayerGO(true, true, true);
+        	var data = pages.nativeAppLayer.getAuthIdText();
+			Console.WriteLine($"AuthIdText: {data}");
+    		Assert.That(data=="AuthId:", Is.True);
+
+			Console.WriteLine("String send: Test");
+			data = altDriver.CallStaticMethod<string>("ConsentManagementProviderLib.CMP", "GetBridgeString", "Assembly-CSharp", new[] { "Test" });
+			Console.WriteLine($"Got: {data}");
+    		Assert.That(data=="Test", Is.True);
+
+			pages.nativeAppLayer.waitForSdkDone();
+			pages.nativeAppLayer.pressClearAll();
+			pages.nativeAppLayer.waitForSdkDone("SDK:Not Started");
+			altDriver.CallStaticMethod<string>("ConsentManagementProviderLib.CMP", "LoadMessage", "Assembly-CSharp", new[] { "AltTesterTest" });
+			pages.nativeAppLayer.waitForSdkDone();
+        	data = pages.nativeAppLayer.getConsentValueText();
+			Console.WriteLine($"ConsentValueText: {data}");
+			if(data=="-")
+			{
+				pages.firstLayerGO(true, true, true);
+				pages.nativeAppLayer.waitForSdkDone();
+			}
+    		Assert.That(data!="-", Is.True);	
 		}
 
         [TearDown]